feat: add blockscope(redundant)
