+++ .gitignore
@@ -61,5 +61,5 @@ gradle-app.setting
 
 # End of https://www.gitignore.io/api/macos,gradle,visualstudiocode
 
-./tests
-./tests/*.sheep
+tests/
+
