+++ src/sheep/chap10/ArrayRunner.java
@@ -5,10 +5,11 @@ import sheep.chap7.ClosureEvaluator;
 import sheep.chap8.NativeEvaluator;
 import sheep.chap9.ClassEvaluator;
 import sheep.chap9.ClassInterpreter;
+import sheep.extra.BlockScopeEvaluator;
 
 public class ArrayRunner {
     public static void main(String[] args) throws Throwable {
         Loader.run(ClassInterpreter.class, args, ClassEvaluator.class, ArrayEvaluator.class, NativeEvaluator.class,
-                ClosureEvaluator.class);
+                ClosureEvaluator.class, BlockScopeEvaluator.class);
     }
 }
