--- tests/LexerRunner.java
@@ -1,11 +1,0 @@
-package tests;
-import sheep.*;
-
-public class LexerRunner {
-    public static void main(String[] args) throws ParseException{
-        Lexer l = new Lexer(new CodeDialog());
-        for (Token t; (t = l.read()) != Token.EOF; ) {
-            System.out.println("=> " + t.getText());
-        }
-    }
-}
