--- tests/plain_test.sheep
@@ -1,5 +1,0 @@
-k = 0
-def hoge() {
-    k = k + 1
-}
-print k
