--- tests/test_script.sheep
@@ -1,26 +1,0 @@
-class Animal {
-    type = "animal"
-    name = "undefined"
-    def getType() {
-        type
-    }
-    def countUp(n) {
-        fun() { n = n+1 }
-    }
-}
-k = 0
-class Sheep extends Animal {
-    name = "sheep"
-    def super_long_long_method_name_for_test_inline_cache() {
-        0
-    }
-}
-
-s = Sheep.new
-
-t1 = currentTime()
-while k < 100000000 {
-    k = k + 1
-    s.super_long_long_method_name_for_test_inline_cache()
-}
-print "time: " + (currentTime() - t1)
